# modified freelist_hmap/hashmapGetFreePageIDs function with better performance