@ahrtr 2023-03-14
superseded by https://github.com/etcd-io/bbolt/pull/419